+++ src/equation/precon.cpp
@@ -1,6 +1,6 @@
 #include "../../include/monolish_blas.hpp"
 #include "../../include/monolish_equation.hpp"
-#include "../monolish_internal.hpp"
+#include "../internal/monolish_internal.hpp"
 
 namespace monolish {