+++ src/utils/coo_util.cpp
@@ -1,10 +1,11 @@
 #include "../../include/common/monolish_dense.hpp"
 #include "../../include/common/monolish_logger.hpp"
 #include "../../include/common/monolish_matrix.hpp"
+#include "../internal/monolish_internal.hpp"
+
 #include <cassert>
 #include <fstream>
 #include <iomanip>
-#include <iostream>
 #include <limits>
 #include <sstream>
 #include <stdexcept>