+++ src/utils/logger_utils.cpp
@@ -1,4 +1,5 @@
 #include "../../include/monolish_blas.hpp"
+#include "../internal/monolish_internal.hpp"
 
 namespace monolish {